backend/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
build==1.3.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
defusedxml==0.7.1
fastapi==0.104.1
filelock==3.18.0
fsspec==2025.7.0
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
h11==0.16.0
hf-xet==1.1.7
httplib2==0.22.0
httptools==0.6.4
huggingface-hub==0.34.4
idna==3.10
packaging==25.0
pip-tools==7.5.0
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.5.0
pydantic_core==2.14.1
pyparsing==3.2.3
pyproject_hooks==1.2.0
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.4
rsa==4.9.1
setuptools==80.9.0
sniffio==1.3.1
starlette==0.27.0
tqdm==4.67.1
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.24.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
wheel==0.45.1
youtube-transcript-api==1.2.2
